update ux for accept bid modal
